refactor id to accessToken for slack

+++ src/server/graphql/models/Integration/integrationMutation.js
@@ -8,7 +8,8 @@ import {
 import {requireSUOrSelf, requireSUOrTeamMember, requireWebsocket} from 'server/utils/authorization';
 import queryIntegrator from 'server/utils/queryIntegrator';
 import addSlackChannel from './addSlackChannel/addSlackChannel';
-
+import {GITHUB} from 'universal/utils/constants';
+import {handleRethinkRemove} from 'server/utils/makeChangefeedHandler';
 export default {
   addSlackChannel,
   removeIntegration: {
@@ -24,7 +25,7 @@ export default {
         description: 'the service to remove'
       }
     },
-    async resolve(source, {teamMemberId, service}, {authToken, socket}) {
+    async resolve(source, {teamMemberId, service}, {authToken, exchange, socket}) {
       const r = getRethink();
 
       // AUTH
@@ -43,13 +44,20 @@ export default {
       });
 
       // TODO refactor projects to hold an id like github's full_name for repos
-      await r.table('Project')
-        .getAll(teamId, {index: 'teamId'})
-        .filter({
-          integrationId: 'foo'
-        })
-        .delete();
-      return res.data;
+      if (service === GITHUB) {
+        await r.table('Project')
+          .getAll(teamId, {index: 'teamId'})
+          .filter({
+            integrationId: 'foo'
+          })
+          .delete();
+      }
+      const oldTokenId = res.data.removeToken;
+      if (oldTokenId) {
+        const payload = handleRethinkRemove({id: oldTokenId});
+        const channel = `integrations/${teamMemberId}`;
+        exchange.publish(channel, payload);
+      }
     }
   }
 };
